requirements.txt
aiogram>=3.4,<4.0
fastapi>=0.110,<1.0
python-dotenv>=1.0,<2.0
gradio_client>=1.3.0,<2.0
pillow>=10.0,<12.0
httpx>=0.27,<1.0
